refactor: update file path for gql-auth.guard
