+++ apps/api/src/modules/provider-chain-members/provider-chain-members.resolver.ts
@@ -3,7 +3,7 @@ import { ProviderChainMember } from './entities/provider-chain-member.entity';
 import { Roles } from 'src/common/decorators/roles.decorator';
 import { UserRoles } from 'src/common/constants/database';
 import { UseGuards } from '@nestjs/common';
-import { GqlAuthGuard } from 'src/common/guards/api-key/gql-auth.guard';
+import { GqlAuthGuard } from 'src/common/guards/gql-auth.guard';
 import { RolesGuard } from 'src/common/guards/role.guard';
 import { QueryOptionsDto } from 'src/common/graphql/dtos/query-options.dto';
 import { ProviderChainMemberResponse } from './dto/provider-chain-member-response.dto';
